+++ mptrack/View_tre.cpp
@@ -4385,6 +4385,9 @@ void CModTree::OnKillFocus(CWnd *pNewWnd)
 	}
 	CTreeCtrl::OnKillFocus(pNewWnd);
 	CMainFrame::GetMainFrame()->m_bModTreeHasFocus = false;
+	// Required to immediately redirect MIDI input focus after drag&drop from tree view to editor
+	if(pNewWnd != nullptr)
+		CMainFrame::GetMainFrame()->SetMidiRecordWnd(pNewWnd->m_hWnd);
 }
 
 
@@ -4396,18 +4399,12 @@ void CModTree::OnSetFocus(CWnd *pOldWnd)
 }
 
 
-bool CModTree::IsItemExpanded(HTREEITEM hItem)
+bool CModTree::IsItemExpanded(HTREEITEM hItem) const
 {
 	// checks if a treeview item is expanded.
-	if(hItem == NULL)
+	if(hItem == nullptr)
 		return false;
-	TVITEM tvi;
-	tvi.mask = TVIF_HANDLE | TVIF_STATE;
-	tvi.state = 0;
-	tvi.stateMask = TVIS_EXPANDED;
-	tvi.hItem = hItem;
-	GetItem(&tvi);
-	return (tvi.state & TVIS_EXPANDED) != 0;
+	return GetItemState(hItem, TVIS_EXPANDED) != 0;
 }