+++ mptrack/dlg_misc.cpp
@@ -755,6 +755,7 @@ void CLegacyPlaybackSettingsDlg::OnFilterStringChanged()
 		case kITCarryAfterNoteOff: desc = _T("Note-Off status does not influence Envelope Carry behaviour"); break;
 		case kFT2OffsetMemoryRequiresNote: desc = _T("Offset effect memory is only updated when the command is next to a note"); break;
 		case kITNoteCutWithPorta: desc = _T("Note Cut (SCx) resets note pitch and interacts with tone portamento + row delay"); break;
+		case kITVolColNoSlidePropagation: desc = _T("Do not propagate volume column volume slide memory to regular effect column"); break;
 
 		default: MPT_ASSERT_NOTREACHED();
 		}
@@ -1387,11 +1388,11 @@ BOOL CEditHistoryDlg::OnInitDialog()
 		CString sDate = CString(_T("<unknown date>"));
 		if(entry.HasValidDate())
 		{
-			const mpt::Date::Unix unixdate = ((m_modDoc.GetSoundFile().GetTimezoneInternal() == mpt::Date::LogicalTimezone::Local) || (m_modDoc.GetSoundFile().GetTimezoneInternal() == mpt::Date::LogicalTimezone::Unspecified))
-				? mpt::Date::UnixFromLocal(mpt::Date::interpret_as_timezone<mpt::Date::LogicalTimezone::Local>(entry.loadDate))
-				: mpt::Date::UnixFromUTC(mpt::Date::interpret_as_timezone<mpt::Date::LogicalTimezone::UTC>(entry.loadDate));
+			const mpt::chrono::default_system_clock::time_point unixdate = ((m_modDoc.GetSoundFile().GetTimezoneInternal() == mpt::Date::LogicalTimezone::Local) || (m_modDoc.GetSoundFile().GetTimezoneInternal() == mpt::Date::LogicalTimezone::Unspecified))
+				? mpt::Date::default_from_local(mpt::Date::interpret_as_timezone<mpt::Date::LogicalTimezone::Local>(entry.loadDate))
+				: mpt::Date::default_from_UTC(mpt::Date::interpret_as_timezone<mpt::Date::LogicalTimezone::UTC>(entry.loadDate));
 				;
-			sDate = CTime(mpt::Date::UnixAsSeconds(unixdate)).Format(_T("%d %b %Y, %H:%M:%S"));
+			sDate = CTime(mpt::chrono::default_system_clock::to_unix_seconds(unixdate)).Format(_T("%d %b %Y, %H:%M:%S"));
 		}
 		// Time + stuff
 		uint32 duration = mpt::saturate_round<uint32>(entry.openTime / HISTORY_TIMER_PRECISION);